requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
beautifulsoup4
blinker
bleach
cachetools
CacheControl
certifi
cffi
chardet
charset-normalizer
click
cleo
colorama
crashtest
cryptography
cssselect
Cython
dataclasses-json
distlib
distro
dnspython
docutils
dulwich
email_validator
exceptiongroup
execnet
fastapi
fastapi-cli
filelock
filetype
frozenlist
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-auth
google-search-results
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
html-to-markdown
html5lib
httpcore
httptools
httpx
httpx-sse
idna
importlib-metadata
importlib-resources
iniconfig
Jinja2
jaraco.classes
jeepney
jpype1
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
keyring
kreuzberg
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-ollama
langchain-text-splitters
langsmith
lockfile
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
more-itertools
msgpack
multidict
mypy_extensions
narwhals
numpy
ollama
orjson
outcome
packaging
pandas
parsel
pdf2image
pexpect
pillow
platformdirs
playa-pdf
pluggy
proto-plus
protobuf
ptyprocess
pyarrow
pyasn1
pyasn1_modules
pycparser
pydeck
Pygments
PyMuPDF
pypdfium2
pyrsistent
pytesseract
PySocks
python-calamine
python-dateutil
python-dotenv
python-multipart
python-pptx
pytz
PyYAML
rapidfuzz
readme-renderer
referencing
reportlab
requests
requests-toolbelt
rfc3986
rich
rich-toolkit
rpds-py
rsa
SecretStorage
selectolax
selenium
selenium-stealth
shellingham
six
smmap
sniffio
sortedcontainers
soupsieve
SQLAlchemy
starlette
streamlit
tabula-py
tenacity
toml
tomli
tomlkit
tornado
tqdm
trio
trio-websocket
trove-classifiers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
virtualenv
w3lib
watchdog
watchfiles
webencodings
websockets
webdriver-manager
wsproto
XlsxWriter
yarl
zipp
zstandard



# aiofiles==24.1.0
# aiohappyeyeballs==2.6.1
# aiohttp==3.11.18
# aiosignal==1.3.2
# altair==5.5.0
# annotated-types==0.7.0
# anyio==4.9.0
# attrs==25.3.0
# beautifulsoup4==4.13.4
# blinker==1.9.0
# cachetools==5.5.2
# certifi==2025.1.31
# chardet==5.2.0
# charset-normalizer==3.4.1
# click==8.1.8
# colorama==0.4.6
# dataclasses-json==0.6.7
# distro==1.9.0
# dnspython==2.7.0
# email_validator==2.2.0
# fastapi==0.115.12
# fastapi-cli==0.0.7
# filetype==1.2.0
# frozenlist==1.6.0
# gitdb==4.0.12
# GitPython==3.1.44
# google-ai-generativelanguage==0.6.17
# google-api-core==2.24.2
# google-auth==2.39.0
# googleapis-common-protos==1.70.0
# greenlet==3.2.1
# grpcio==1.71.0
# grpcio-status==1.71.0
# gunicorn==23.0.0
# h11==0.14.0
# html-to-markdown==1.3.1
# httpcore==1.0.8
# httptools==0.6.4
# httpx==0.28.1
# httpx-sse==0.4.0
# idna==3.10
# Jinja2==3.1.6
# jpype1==1.5.2
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# kreuzberg==3.1.3
# langchain==0.3.24
# langchain-community==0.3.22
# langchain-core==0.3.55
# langchain-google-genai==2.1.3
# langchain-ollama==0.3.2
# langchain-text-splitters==0.3.8
# langsmith==0.3.33
# lxml==5.3.2
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# marshmallow==3.26.1
# mdurl==0.1.2
# multidict==6.4.3
# mypy_extensions==1.1.0
# narwhals==1.35.0
# numpy==2.2.5
# ollama==0.4.8
# orjson==3.10.16
# packaging==24.2
# pandas==2.2.3
# pdf2image==1.17.0
# pillow==11.2.1
# playa-pdf==0.4.1
# propcache==0.3.1
# proto-plus==1.26.1
# protobuf==5.29.4
# pyarrow==19.0.1
# pyasn1==0.6.1
# pyasn1_modules==0.4.2
# pydantic==2.11.3
# pydantic-settings==2.9.1
# pydantic_core==2.33.1
# pydeck==0.9.1
# Pygments==2.19.1
# PyMuPDF==1.25.5
# pypdfium2==4.30.0
# pytesseract==0.3.13
# python-calamine==0.3.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.0
# python-multipart==0.0.20
# python-pptx==1.0.2
# pytz==2025.2
# PyYAML==6.0.2
# referencing==0.36.2
# reportlab==4.4.0
# requests==2.32.3
# requests-toolbelt==1.0.0
# rich==14.0.0
# rich-toolkit==0.14.1
# rpds-py==0.24.0
# rsa==4.9.1
# shellingham==1.5.4
# six==1.17.0
# smmap==5.0.2
# sniffio==1.3.1
# soupsieve==2.7
# SQLAlchemy==2.0.40
# starlette==0.46.2
# streamlit==1.44.1
# tabula-py==2.10.0
# tenacity==9.1.2
# toml==0.10.2
# tornado==6.4.2
# typer==0.15.2
# typing-inspect==0.9.0
# typing-inspection==0.4.0
# typing_extensions==4.13.2
# tzdata==2025.2
# urllib3==2.4.0
# uvicorn==0.34.2
# watchdog==6.0.0
# watchfiles==1.0.5
# websockets==15.0.1
# XlsxWriter==3.2.3
# yarl==1.20.0
# zstandard==0.23.0

# # for google scholar
# async-generator==1.10
# # attrs==22.2.0
# bleach==6.0.0
# CacheControl==0.12.11
# certifi==2022.12.7
# cffi==1.15.1
# charset-normalizer==3.0.1
# cleo==2.0.1
# colorama==0.4.6
# crashtest==0.4.1
# cryptography==39.0.1
# cssselect==1.2.0
# Cython==0.29.33
# distlib==0.3.6
# docutils==0.19
# dulwich==0.20.50
# exceptiongroup==1.1.0
# execnet==1.9.0
# filelock==3.9.0
# google-search-results==2.4.2
# h11==0.14.0
# html5lib==1.1
# idna==3.4
# importlib-metadata==4.13.0
# importlib-resources==5.10.2
# iniconfig==2.0.0
# jaraco.classes==3.2.3
# jeepney==0.8.0
# jsonschema==4.17.3
# keyring==23.13.1
# lockfile==0.12.2
# lxml==5.4.0
# markdown-it-py==2.1.0
# mdurl==0.1.2
# more-itertools==9.0.0
# msgpack==1.0.4
# # numpy==1.24.2
# outcome==1.2.0
# packaging==23.0
# pandas==1.5.3
# parsel==1.7.0
# pexpect==4.8.0
# pkginfo==1.9.6
# pkgutil_resolve_name==1.3.10
# platformdirs==2.6.2
# pluggy==1.0.0
# poetry==1.3.2
# poetry-core==1.4.0
# poetry-plugin-export==1.3.0
# ptyprocess==0.7.0
# pycparser==2.21
# Pygments==2.14.0
# pyrsistent==0.19.3
# PySocks==1.7.1
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# pytz==2022.7.1
# pywin32-ctypes==0.2.0
# rapidfuzz==2.13.7
# readme-renderer==37.3
# requests==2.28.2
# requests-toolbelt==0.10.1
# rfc3986==2.0.0
# rich==13.3.1
# SecretStorage==3.3.3
# selectolax==0.3.28
# selenium==4.8.0
# selenium-stealth==1.0.6
# shellingham==1.5.0.post1
# six==1.16.0
# sniffio==1.3.0
# sortedcontainers==2.4.0
# tomli==2.0.1
# tomlkit==0.11.6
# tqdm==4.65.0
# trio==0.22.0
# trio-websocket==0.9.2
# trove-classifiers==2023.1.20
# typing_extensions==4.4.0
# urllib3==1.26.14
# virtualenv==20.19.0
# w3lib==2.1.1
# webdriver-manager==3.8.5
# webencodings==0.5.1
# wsproto==1.2.0
# zipp==3.12.1
